+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#184](https://github.com/babylonlabs-io/finality-provider/pull/184) eots manager sign record store
 * [#189](https://github.com/babylonlabs-io/finality-provider/pull/189) Remove `fpd register-finality-provider` cmd
 * [#190](https://github.com/babylonlabs-io/finality-provider/pull/190) Benchmark pub rand
+* [#193](https://github.com/babylonlabs-io/finality-provider/pull/193) adds unsafeSignEOTS for e2e tests
 * [#195](https://github.com/babylonlabs-io/finality-provider/pull/195) Not block unjailing
 
 ### Bug Fixes